Add authentication filter code
An envoy filter to orchestrate authentication with some basic token
(i.e. API key). This filter manipulates the request headers - injecting
a JWT obtained from some upstream. The JWT may then be passed forward
toward the intended destination, possibly through other filters that
verify the JWT and/or use the claims with in.