Add integration tests
Integration tests are achieved through a Go package orchestrating a
local version of the envoy proxy. Before each test scenario envoy is
configured to use the compiled WASM filter, and is exposed on a local
port. Then the scenario is run by making a request to envoy and the
response asserted.

The Go package requires that the `bazel` command is present for it to
run the `Envoy` step. However it is not required for the application to
be built, nor the tests run, using Bazel. As such it is enough to only
include the `bazel-bin` directory and an empty `WORKSPACE` file at the
root of the application.